Moved `get_component(_unchecked_mut)` from `Query` to `QueryState`
Also had to move `QueryComponentError` to match the new location of its methods.

Also performed a QA pass to narrow the scope of `unsafe` blocks within `Query` methods.
